[Java] Fix a bug in IntHashSet's retainAll implementation where it was possible for the new capacity that is used in a rehash operation to be the same as the new size of the set, causing a .contains() call to loop infinitely.
The new capacity was being calculated by finding the next power of 2 or the new size itself if it is a power of 2.

When the new size of the set is a power of two there will not be a space for the sentinel missing value.

Fixed by choosing the next power of two when this happens.