BroadcastReceiver lapCount opaque access.
The BroadcastReceiver makes use of opaque access for the lappedCount.

LappedCount is a progress indicator and doesn't need to provide any
ordering guaruantees for surrounding loads/stores. The only thing
that needs to be guarantees is that the value is read atomically
and that once you read value, on a subsequent read, you can't read
an earlier version (coherence).