CountersManager setCounterOwnerId opaque write.
A plain write was done but the compiler could play with that
write and delay its visibility. Better to make use of an opaque
write to ensure that the compiler will emit the store.

I checked the usage e.g. the DriverConductor:

{
   final AeronClient client = getOrAddClient(clientId);
   final AtomicCounter counter = countersManager.newCounter(
            typeId, keyBuffer, keyOffset, keyLength, labelBuffer, labelOffset, labelLength);

   countersManager.setCounterRegistrationId(counter.id(), correlationId);
   countersManager.setCounterOwnerId(counter.id(), clientId);
   counterLinks.add(new CounterLink(counter, correlationId, client));
   clientProxy.onCounterReady(correlationId, counter.id());
}

And I don't see any obvious barrier behavior that will ensure that the write
to the owner id is made.