Fixes 2 bugs in the BufferUtil boundCheck. (#347)
If the index is close to Long.MAX_VALUE, then the resulting pos can becomes negative
due to overflow.

As a consequence, the resultingPos > capacity will not be triggered even though
the resultingPos has exceeded the capacity of the buffer.

In the Agrona codebase this has never lead to problems because only
int values are passed to index and therefor there won't be any overflow.

Another change is it length is also checked for being negative.